chore(executor): extract explain.rs (#116 step 5)
Pulls the tri-dialect EXPLAIN machinery (issue #272 / T1.10) into
explain.rs (241 LOC):

- `ExplainOptions` knob struct (analyze / buffers / verbose / format).
- `ExplainFormat` enum (Text / Json / Yaml / Xml).
- `impl ExplainOptions::to_clause` — renders `(ANALYZE, BUFFERS,
  FORMAT JSON)` parenthesized PG clause.
- `explain_pool` — tri-dialect dispatcher that emits PG's
  `EXPLAIN (FORMAT ...) <stmt>`, MySQL's `EXPLAIN [ANALYZE]
  [FORMAT=JSON|TREE] <stmt>`, or SQLite's `EXPLAIN QUERY PLAN <stmt>`.

The `QuerySet::explain` / `QuerySet::explain_on` inherent methods
stay in mod.rs since they're impls on the QuerySet inherent surface;
they consume `ExplainOptions` / `ExplainFormat` re-exported from
the new module.

mod.rs: 3852 → 3640 LOC (-212). Build clean across sqlite,tenancy
litmus + all-features. 1482 lib tests pass.
